makemake.pl: delete the files mesh_params.h and unrolled_loops.f90, if they exist. Otherwise they sometimes mess up the Makefile (makemake.pl includes EVERY .f90 and .h file into Makefile, even those which do not belong to the program. Which is a mess.)
submit.csh: The idea of make is to check whether any source files are newer than its dependants and if so, recompile them. The submit script does not have to replicate this functions. So it now runs 'make all', checks the return code, if it is zero, continues and otherwise exits.

The same goes for the submit script in the SOLVER dir.

The old approach had bazillions of possible loopholes (only one source file changed. or: make failed, but xsem was still existing from previous run, or...)